+++ src/index.tsx
@@ -1,7 +1,7 @@
 import { ApolloProvider } from '@apollo/client';
 import ReactDOM from 'react-dom';
 import App from './App';
-import client from './client';
+import client from './apollo';
 import './tailwind.css';
 
 ReactDOM.render(
